+++ include/partition.h
@@ -27,7 +27,7 @@
     RETURN
     Pivot index in t
 */
-size_t partition_get_pivot_trivial(const void *t, size_t len, int size_of, int (*cmp)(const void *a, const void *b));
+size_t partition_get_pivot_trivial(const void *t, size_t len, size_t size_of, int (*cmp)(const void *a, const void *b));
 
 /*
     Get index of pivot using rand
@@ -41,7 +41,7 @@ size_t partition_get_pivot_trivial(const void *t, size_t len, int size_of, int (
     RETURN
     Pivot index in t
 */
-size_t partition_get_pivot_rand(const void *t, size_t len, int size_of, int (*cmp)(const void *a, const void *b));
+size_t partition_get_pivot_rand(const void *t, size_t len, size_t size_of, int (*cmp)(const void *a, const void *b));
 
 /*
     Get index of pivot using tukey heuristic (medians of 9)
@@ -55,7 +55,7 @@ size_t partition_get_pivot_rand(const void *t, size_t len, int size_of, int (*cm
     RETURN
     Pivot index in t
 */
-size_t partition_get_pivot_median(const void *t, size_t len, int size_of, int (*cmp)(const void *a, const void *b));
+size_t partition_get_pivot_median(const void *t, size_t len, size_t size_of, int (*cmp)(const void *a, const void *b));
 
 /*
     Bentley McIlory partition
@@ -77,7 +77,7 @@ ssize_t partition_bentley(void        *t,
                           size_t      offset_left,
                           size_t      offset_right,
                           int         (*cmp)(const void *a, const void *b),
-                          int         size_of,
+                          size_t      size_of,
                           size_t      *offset_left_index,
                           size_t      *offset_right_index);
 
@@ -101,7 +101,7 @@ ssize_t partition_hoare(void        *t,
                         size_t      offset_left,
                         size_t      offset_right,
                         int         (*cmp)(const void *a, const void *b),
-                        int         size_of,
+                        size_t      size_of,
                         size_t      *offset_left_index,
                         size_t      *offset_right_index);
 
@@ -125,7 +125,7 @@ ssize_t partition_lomuto(void        *t,
                          size_t      offset_left,
                          size_t      offset_right,
                          int         (*cmp)(const void *a, const void *b),
-                         int         size_of,
+                         size_t      size_of,
                          size_t      *offset_left_index,
                          size_t      *offset_right_index);
