+++ include/ulist.h
@@ -62,7 +62,7 @@ typedef struct UList
     int         (*____get_pos)(const void *list, size_t pos, void *data);
     void*       (*____merge)(const void * ___restrict___ list1, const void * ___restrict___ list2);
     int         (*____to_array)(const void *list, void *array, size_t *size);
-    int         (*____get_data_size)(const void *list);
+    ssize_t     (*____get_data_size)(const void *list);
     ssize_t     (*____get_num_entries)(const void *list);
 
     /* to create iterator */
@@ -121,7 +121,7 @@ ___inline___ int ulist_iterator_init(const UList *list, UList_iterator *it, iti_
     static ___unused___ int ____get_pos(const void *list, size_t pos, void *data); \
     static ___unused___ void *____merge(const void * ___restrict___ list1, const void * ___restrict___ list2); \
     static ___unused___ int ____to_array(const void *list, void *array, size_t *size); \
-    static ___unused___ int ____get_data_size(const void *list); \
+    static ___unused___ ssize_t ____get_data_size(const void *list); \
     static ___unused___ ssize_t ____get_num_entries(const void *list); \
     static ___unused___ void ____destroy(void *list) \
     { \
@@ -192,7 +192,7 @@ ___inline___ int ulist_iterator_init(const UList *list, UList_iterator *it, iti_
         return concat(prefix, _to_array)((const type *)list, array, size); \
     } \
     \
-    static ___unused___ int ____get_data_size(const void *list) \
+    static ___unused___ ssize_t ____get_data_size(const void *list) \
     { \
         return concat(prefix, _get_data_size)((const type *)list); \
     } \
@@ -447,7 +447,7 @@ ___inline___ int ulist_to_array(const UList *list, void *array, size_t *size);
     -1 iff failure
     UList data size of iff success
 */
-___inline___ int ulist_get_data_size(const UList *list);
+___inline___ ssize_t ulist_get_data_size(const UList *list);
 
 /*
     Get UList num of entries
@@ -661,7 +661,7 @@ ___inline___ int ulist_to_array(const UList *list, void *array, size_t *size)
     return list->____to_array(ulist_get_list(list), array, size);
 }
 
-___inline___ int ulist_get_data_size(const UList *list)
+___inline___ ssize_t ulist_get_data_size(const UList *list)
 {
     if (list == NULL)
         return 1;
